refactor(drive): split executors.rs into one file per DocumentCountMode variant
`executors.rs` had one `impl Drive { ... }` block holding all
six per-mode executor methods (plus a private helper) at ~480
LoC. Each method is independent of the others — they share no
state, only the `impl Drive` block. Splitting along mode
boundaries makes "find the X-mode executor" a directory lookup
rather than a grep through 480 lines.

New layout:

    drive_document_count_query/
    └── executors/
        ├── mod.rs                    (32 LoC, module decls + overview)
        ├── total.rs                  (114 LoC) — DocumentCountMode::Total + read_primary_key_count_tree helper
        ├── per_in_value.rs           (158 LoC) — DocumentCountMode::PerInValue
        ├── range_no_proof.rs         (54 LoC)  — DocumentCountMode::RangeNoProof
        ├── range_proof.rs            (50 LoC)  — DocumentCountMode::RangeProof
        ├── range_distinct_proof.rs   (70 LoC)  — DocumentCountMode::RangeDistinctProof
        └── point_lookup_proof.rs     (91 LoC)  — DocumentCountMode::PointLookupProof

The `read_primary_key_count_tree` private helper stays in
`total.rs` (its only caller); it's `pub(super)` so a future
PointLookupProof variant that needs it can pull it in without
re-declaring.

`executors.rs` file deleted, replaced by the directory module.
No re-exports needed — each file adds methods directly to
`impl Drive`, so the dispatcher sees them on the `Drive` type
unchanged.

39 + 51 + 7 tests still pass.
